moved symmetric and Asymmetric to separate files
